ACCUMULO-2589 Drop deprecated code from 1.x

+++ core/src/main/java/org/apache/accumulo/core/client/impl/ThriftScanner.java
@@ -213,7 +213,7 @@ public class ThriftScanner {
     return (long) (Math.max(millis * 2, 3000) * (.9 + Math.random() / 5));
   }
 
-  public static List<KeyValue> scan(ClientContext context, ScanState scanState, int timeOut) throws ScanTimedOutException, AccumuloException,
+  public static List<KeyValue> scan(ClientContext context, ScanState scanState, long timeOut) throws ScanTimedOutException, AccumuloException,
       AccumuloSecurityException, TableNotFoundException {
     TabletLocation loc = null;
     Instance instance = context.getInstance();
